backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
langchain==0.2.5
langchain-community==0.2.5
langchain-google-genai==1.0.5
chromadb==0.5.0
numpy==1.26.4
Flask-Cors==4.0.1
requests==2.32.3
werkzeug==3.0.3 